requirements.txt
flask
flask-cors
sqlalchemy
python-dotenv
twilio
pinecone-client
psycopg2-binary
r2r
sentence-transformers
PyPDF2
python-docx
pandas
openai
litellm
httpx
aiofiles